legacy-python/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-genai>=1.60.0
msal>=1.24.0
openpyxl>=3.1.0
openai>=1.0.0
anthropic>=0.18.0
flask>=3.0.0
gunicorn>=21.2.0
Flask-Session>=0.5.0


